splits feedreader cmd in 2, for articles and trials
